corrigiendo cursos usando sweet alert y token
validando de que el dato que existe tire un mensaje bajo el input, de que ese dato ya existe

+++ src/app/api/courses.api.ts
@@ -17,8 +17,12 @@ export async function getAllCourses(
   return await response.json();
 }
 
-export async function addCourse(courseData: CoursesData) {
-  const token = "";
+export async function addCourse(
+  courseData: CoursesData,
+  token: string | undefined
+) {
+  if (!token) throw new Error("Token no encontrado");
+
   const res = await fetch("http://localhost:4000/api/v1/courses", {
     method: "POST",
     headers: {
@@ -27,11 +31,25 @@ export async function addCourse(courseData: CoursesData) {
     },
     body: JSON.stringify(courseData),
   });
-  return res.json();
+
+  const data = await res.json();
+
+  if (!res.ok) {
+    const error = new Error(data.message || "Error al agregar el curso") as any;
+    error.response = { status: res.status, data };
+    throw error;
+  }
+
+  return data;
 }
 
-export async function updateCourse(courseData: CoursesData, id: number) {
-  const token = "";
+export async function updateCourse(
+  courseData: CoursesData,
+  id: number,
+  token: string | undefined
+) {
+  if (!token) throw new Error("Token no encontrado");
+
   const res = await fetch(`http://localhost:4000/api/v1/courses/${id}`, {
     method: "PUT",
     headers: {
@@ -40,11 +58,22 @@ export async function updateCourse(courseData: CoursesData, id: number) {
     },
     body: JSON.stringify(courseData),
   });
-  return res.json();
+
+  const data = await res.json();
+
+  if (!res.ok) {
+    const error = new Error(
+      data.message || "Error al actualizar el curso"
+    ) as any;
+    error.response = { status: res.status, data };
+    throw error;
+  }
+
+  return data;
 }
 
-export async function deleteCourse(id: number) {
-  const token = "";
+export async function deleteCourse(id: number, token: string | undefined) {
+  if (!token) throw new Error("Token no encontrado");
   const res = await fetch(`http://localhost:4000/api/v1/courses/${id}`, {
     method: "DELETE",
     headers: {
